move KubernetesVersionProvider to its own package (#901)
