+++ cmd/controller/main_ccp.go
@@ -75,8 +75,8 @@ func main() {
 			op.EventRecorder,
 			aksCloudProvider,
 			op.InstanceProvider,
-			// TODO: refactor so we are passing in kubernetesVersionProvider and nodeImageProvider. Currently ImageProvider just implements both.
-			op.ImageProvider,
+			// TODO: still need to refactor ImageProvider side of things.
+			op.KubernetesVersionProvider,
 			op.ImageProvider,
 			op.InClusterKubernetesInterface,
 		)...).
