+++ pkg/providers/kubernetesversion/kubernetesversion.go
@@ -14,8 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// TODO (charliedmcb): factor provider into its own package
-package imagefamily
+package kubernetesversion
 
 import (
 	"context"
